Eliminate duplicated page object (CoursesListPage), use Dashboard instead of it.
refs: MBL-16376
affects: Teacher
release note: none
