+++ apps/teacher/src/androidTest/java/com/instructure/teacher/ui/AnnouncementsListPageTest.kt
@@ -153,7 +153,7 @@ class AnnouncementsListPageTest : TeacherTest() {
         val token = data.tokenFor(teacher)!!
         tokenLogin(data.domain, token, teacher)
 
-        coursesListPage.openCourse(course)
+        dashboardPage.openCourse(course)
         courseBrowserPage.openAnnouncementsTab()
         return data
     }
