+++ apps/teacher/src/androidTest/java/com/instructure/teacher/ui/AssigneeListPageTest.kt
@@ -92,7 +92,7 @@ class AssigneeListPageTest : TeacherTest() {
         val token = data.tokenFor(teacher)!!
         tokenLogin(data.domain, token, teacher)
 
-        coursesListPage.openCourse(course)
+        dashboardPage.openCourse(course)
         courseBrowserPage.openAssignmentsTab()
         assignmentListPage.clickAssignment(assignment)
