+++ apps/teacher/src/androidTest/java/com/instructure/teacher/ui/AssignmentListPageTest.kt
@@ -96,7 +96,7 @@ class AssignmentListPageTest : TeacherTest() {
         val token = data.tokenFor(teacher)!!
         tokenLogin(data.domain, token, teacher)
 
-        coursesListPage.openCourse(course)
+        dashboardPage.openCourse(course)
         courseBrowserPage.openAssignmentsTab()
         return data
     }
