+++ apps/teacher/src/androidTest/java/com/instructure/teacher/ui/CommentLibraryPageTest.kt
@@ -246,7 +246,7 @@ class CommentLibraryPageTest : TeacherTest() {
         val token = data.tokenFor(teacher)!!
         tokenLogin(data.domain, token, teacher)
 
-        coursesListPage.openCourse(course)
+        dashboardPage.openCourse(course)
         courseBrowserPage.openAssignmentsTab()
         assignmentListPage.clickAssignment(assignment)
         assignmentDetailsPage.openSubmissionsPage()
