+++ apps/teacher/src/androidTest/java/com/instructure/teacher/ui/DiscussionsListPageTest.kt
@@ -78,7 +78,7 @@ class DiscussionsListPageTest : TeacherTest() {
         val token = data.tokenFor(teacher)!!
         tokenLogin(data.domain, token, teacher)
 
-        coursesListPage.openCourse(course)
+        dashboardPage.openCourse(course)
         courseBrowserPage.openDiscussionsTab()
         return data
     }
