+++ apps/teacher/src/androidTest/java/com/instructure/teacher/ui/EditQuizDetailsPageTest.kt
@@ -194,7 +194,7 @@ class EditQuizDetailsPageTest : TeacherTest() {
         val token = data.tokenFor(teacher)!!
         tokenLogin(data.domain, token, teacher)
 
-        coursesListPage.openCourse(course)
+        dashboardPage.openCourse(course)
         courseBrowserPage.openQuizzesTab()
         quizListPage.clickQuiz(quiz)
         quizDetailsPage.openEditPage()
