+++ apps/teacher/src/androidTest/java/com/instructure/teacher/ui/InboxMessagePageTest.kt
@@ -19,7 +19,9 @@ package com.instructure.teacher.ui
 import com.instructure.canvas.espresso.mockCanvas.MockCanvas
 import com.instructure.canvas.espresso.mockCanvas.addConversations
 import com.instructure.canvas.espresso.mockCanvas.init
-import com.instructure.teacher.ui.utils.*
+import com.instructure.teacher.ui.utils.TeacherTest
+import com.instructure.teacher.ui.utils.clickInboxTab
+import com.instructure.teacher.ui.utils.tokenLogin
 import dagger.hilt.android.testing.HiltAndroidTest
 import org.junit.Test
 
@@ -51,7 +53,7 @@ class InboxMessagePageTest: TeacherTest() {
 
         val token = data.tokenFor(teacher)!!
         tokenLogin(data.domain, token, teacher)
-        coursesListPage.clickInboxTab()
+        dashboardPage.clickInboxTab()
         inboxPage.clickConversation(chosenConversation)
     }
 }
