+++ apps/teacher/src/androidTest/java/com/instructure/teacher/ui/InboxPageTest.kt
@@ -20,7 +20,9 @@ package com.instructure.teacher.ui
 import com.instructure.canvas.espresso.mockCanvas.MockCanvas
 import com.instructure.canvas.espresso.mockCanvas.addConversations
 import com.instructure.canvas.espresso.mockCanvas.init
-import com.instructure.teacher.ui.utils.*
+import com.instructure.teacher.ui.utils.TeacherTest
+import com.instructure.teacher.ui.utils.clickInboxTab
+import com.instructure.teacher.ui.utils.tokenLogin
 import dagger.hilt.android.testing.HiltAndroidTest
 import org.junit.Test
 
@@ -37,7 +39,7 @@ class InboxPageTest: TeacherTest() {
         data.addConversations(userId = teacher.id)
         val token = data.tokenFor(teacher)!!
         tokenLogin(data.domain, token, teacher)
-        coursesListPage.clickInboxTab()
+        dashboardPage.clickInboxTab()
         inboxPage.assertPageObjects()
     }
 
@@ -58,7 +60,7 @@ class InboxPageTest: TeacherTest() {
 
         val token = data.tokenFor(teacher)!!
         tokenLogin(data.domain, token, teacher)
-        coursesListPage.clickInboxTab()
+        dashboardPage.clickInboxTab()
         inboxPage.assertHasConversation()
     }
 }
