+++ apps/teacher/src/androidTest/java/com/instructure/teacher/ui/PageListPageTest.kt
@@ -71,7 +71,7 @@ class PageListPageTest : TeacherTest() {
 
         val token = data.tokenFor(teacher)!!
         tokenLogin(data.domain, token, teacher)
-        coursesListPage.openCourse(course)
+        dashboardPage.openCourse(course)
         courseBrowserPage.openPagesTab()
         return pages
     }
